Add missing newlines in docs pages (#4108)
* Add missing newline in users-guide/09-pulses.md

Missing newline causes text rendered on metabase.com docs site to be in a paragraph instead of a bulleted list. It looks fine on GitHub though.

* Add missing newline in admin-guide/03-metadata.md

Missing newline causes text rendered on metabase.com docs site to be in a paragraph instead of a bulleted list. It looks fine on GitHub though.

* Fix link in admin-guide/10-sso

Link at the bottom of the page pointed to old number for the getting-started-guide