Enable foreign-keys for Drill and Spark SQL
The tests don't work since we can't create foreign keys with those
SQL engines, but they can perform joins if the foreign key has been
specified manually in the data model interface.

We should not include the schema name in the "ON" clause (it breaks
Drill). Note that the schema is already correctly specified in the
"FROM <schema>.table1" and "LEFT JOIN <schema>.table2 alias1" clauses,
so there should be no ambiguity in in "ON table1.field1 = alias.field2"