Add support for remapping user provided enum-like values
This commit adds storage of dimension data that the user provides for
an existing field. This is to translate enum-like
fields (i.e. integers) to a textual description of that data. This
remapped value will be included in the response with the original
value. The UI will then take that and show the more user friendly
remapped value.

The new "remapped_from" and "remapped_to" attributes in the cols field
of the response will indicate that a field replaces another field.