Add FK related remapping
This commit adds logic around an FK remapping found in the
dimensions table. It will follow a foreign key to the related
table and use the indicated column from that table as the remapped
value for the source column.

This commit also switches the "remapped_from" and "remapped_to"
attributes returned in the query response to use the "name" of the
remapped column as there could be ambiguity with the previously used
display name.