java.jdbc column renames should update metadata
Previously when java.jdbc renamed a project column (i.e. NAME to
NAME_2) to disambiguate them, we would discard previous information on
that NAME column and reconstruct it based on the data that comes back
from the query. This changes that behavior to update the existing
column data to the new name. When we do not have metadata on that
column it will use the existing behavior to infer that.

This commit also adds some tests for remapping which uncovered this
bug.