Add the "bin-width" binning strategy
In MBQL you can now specify a bin-width binning strategy with a
specified bin-width and the BinnedField will bin using that
information.

This commit also changes the existing behavior of the "default"
binning strategy. When default is specified (for now) this uses the
"num-bins" strategy with the globally set default number of
bins (specified via setting). Specifying num-bins or bin-width
requires an extra parameter specifying the number of bins or the bin
width. Only default uses the global setting.