Clear dimension data when the special_type is not supported
If a field is set to an enum or category type and has a remapping,
that should be cleared if the field moves to a special_type that is
not allowed to have remappings (i.e. text). This commit adds checks
for that and clears the dimension in that case.