Fixup human readable values when field values change [ci drivers]
Field values and human readable values are zipped together to support
remapping. If the field values change (i.e. new values, removing
values etc) then the pairings of those field values and their remapped
values can be wrong. This commit adds some code to reconcile that. If
a new field value is found, a stringified version of that field value
is used as the human readable version (the user can update/change this
later).

Fixes #5782