Ensure timezone detection doesn't cause sync to fail [ci drivers]
This commit wraps the timezone detections stuff in some try/catches so
that if there's an issue parsing one of the timestamps from the
database, it doesn't cause the whole sync process to fail.

If a failure happens, it's logged as a warning, the timezone field in
the database is not updated and the sync process proceeds.

Fixes #5869