Fix a vertica test race condition in the unix timestamp tests [ci drivers]
The problem appears to be reusing a connection with a timezone set for
that session from a previous test that was ran. The report timezone is
not specified, but the timezone is set on the session from a previous
run. This results in the timestamps being return as UTC, but shifted
back 7 hours (as if it were pacific time).

This commit just specified a report-timezone of UTC to avoid that issue.