Fix issue with scalar pulses returning timezones
This commit fixes a format-cell argument order issue that would cause
the timezone object to be returned instead of the data being
formatted.

While investigating this I found that the database_id isn't being
populated and so we are erroring out while trying to lookup the report
timezone. This commit will look in the database_query map if
database_id isn't found on the card.