Add the OpneSourceSubProtocolOverride flag to PG and Redshift [ci drivers]
Redshift JDBC connection URLs can be writtend with the postgresql
subprotocol or the redshift protocol. This can cause problems as the
DriverManager (depending on how the classes have been loaded) will be
the first matching driver for a subprotocol. We could get the Redshift
driver when we were intending to get the PostgreSQL driver. Adding
this flag will ensure that we don't inadvertantly get the Redshift
driver when we intended to get PostgreSQL driver.