Add an in-context test fixture to help diagnose stale test data
This commit adds an expectation_options namespace that the
expectations library will automatically load for functions that should
run before/after/around each test expectation.

An 'in-context' is in that namespace that will run around each test
expectations checking for instances of a specified model. If one is
found the test run exits with helpful information in tracking down
where the stale data came from.

By default this function is noop and no models are checked. Adding
modes to the `models-to-check` var will enable this debugging and if
stale data is founc, the `with-model-cleanup` macro in
test/metabase/test/util can be used to ensure that stale data gets
removed.