Alerts Backend
This commit has the backend code needed to support alerts. From the
frontend perspective, Alerts is separate from pulses and has a
different endpoint, however much of the backend code and database
tables are shared. A few new columns were added to the pulse table to
support alerts, but most of the data is the same between pulses and
alerts.

This also adds several "meta-notifications" around alerts to let users
know when they've been added to an alert, removed from an alert or if
a question has changed such that the alert is no longer meaningful.

This fixes #1372, fixes #6261 and fixes #5768.