Improve memory usage of infer-spaces
This commit changes how the word frequency and cost data is
represented in memory. Previously this data consumed about 7 MB on all
Metabase instances. This commit lowers that memory to roughly 1.5
MB. The primary change is storing the hash of the strings in a
primitive integer array instead of the strings themselves. It also
stores the cost in a correlated primitive double array, saving space
there as well.