Move google utilities out of the driver namespace
The metabase.driver.google namespace isn't a driver but rather a set
of utilities used by the Google Analytics and Bigquery driver. When we
try to load the google namespace as a driver it results in a warning
that there is no `-init-driver` function. It's harmless, but not
necessary. This commit moves the code to a utility namespace which
will not try to load it as a driver.