Switch alert-test `expect-with-temp` calls to `expect`
Using `expect-with-temp` rather than `(expect ... (with-temp*` can
obscure some test failures. When using data from the `with-temp*` in
the left side of the `expect` it's needed, but if not, it's better to
nest he `with-temp*` so as to not interfere with the error handling
and test failure reporting.

This commit flips a bunch of the alerts tests that don't need `expect-with-temp`.