Fix async-test transient failures
Depending on the order in which the threads were running, we were
looking for the job in the running-jobs atom BEFORE we put the data in
the atom. If that happens, we might not ever update the job with the
correct status, it's also possible that we'll never remove the job
from the running-jobs atom, just depending on unlucky we are with the
ordering of the threads.

This commit introduces coordination between the code adding the new
data to the running-jobs atom and the work done by the job.