Write JSON to a piped input stream for /database/:id/fields
Previously we would use the default behavior, which is
`ring.middleware.json/wrap-json-response`. `wrap-json-response` will
generate a string from the Clojure data structure and use that as the
ring response. This commit switches serializing to a
`PipedInputStream` which allows us to use less memory and return the
results as we are serializing them to JSON.