Avoid issue metadata updates in sync unless there's a change
Previously every metabase field was updated during sync, giving it a
new updated_at value even if there was no change. This commit will
only update a row if there's been a base_type change or a new special
type was found (which should be pretty rare). This significantly
increases the performance of sync, especially on installs with larger
numbers of columns being sync'd.