Stream all JSON responses [ci drivers]
This commit switches from `ring.middleware.json/wrap-json-response` to
a new function that will stream JSON responses. Before this change the
full JSON response was serialized to a String first, then written to
the response stream. This change will skip that step and write the
response to a PipedOutputStream directly.

This is the same approach that was used in the response for the
`/database/:id/fields` route, just generalized and applied to all
routes.