Stream results from the database to JSON
This commit uses reducible query results from clojure.java.jdbc to
stream application database query results from to the response as
JSON. Previously we streamed the JSON serialization part, but not the
query portion.

This commit is more intended as an RFC and only switches over the
`/database/:id/fields` endpoint to use this. Testing on my machine, on
large synthetic schema, shows it uses about half the memory that the
previous code used and executes in roughly the same wall clock time.

This commit relies on related query-reducible functionality in Toucan
that hasn't been merged yet, hence the dependency on Toucan
1.1.4-SNAPSHOT.