Switch sql templates to an instaparse grammar [ci drivers]
This commit switches the SQL template parsing to a formally specified
EBNF grammar using instaparse. By specifying it in this way, changes
to the language become more flexible and explicit. This commit also
adds support for multiple parameters specified in an optional clause
as it's baked into the grammar.

Fixes #5492