Update pulse code to allow user-specified x and y axes
The previous code had two issues. The first was a bug in using the
user-specified y-axis column for checking if a goal had been
met. Beyond that the pulse code didn't recognize those user specified
x and y axes either. This commit uses the user specified columns, when
they are not present defaults to the previous first/second column
behavior in the resultset.