Add CSV/XLS Pulse attachment support to the backend
This commit adds the backend pieces needed to attach CSV and XLS files
of query results to Pulses. Attachment support is also there for
Alerts. This is an opt-in feature for pulses and alerts. Users can
update existing Pulses to include attachments or add them to a new
pulse. The file(s) will only be attached when there are results.

Fixes #6587, fixes #3894, fixes #2623