Add backend support for time fields
This commit adds code needed to query time fields and to filter by
time fields. Support is included for all database backends that
support time fields. Previously depending on the backend, the result
of querying a time field was either an exception or an incorrect
"zeroed out" date of 1970-01-01 followed by the time.

Fixes #5563, fixes #5829