Update date formatting export code for SQLite
SQLite doesn't have proper datetime support. As a result of this, the
datetimes returned from SQLite pass through the QP untouched, meaning
the datetime format is different than the other drivers. As a result,
exporting a date field to CSV from SQLite would cause a failure.

This commit gets more defensive/flexible in what kind of datetime it
will convert to a date. It will ensure what it's converting is a
string and then look for a "T" (most drivers) or a " " in the
datetime (SQLite). If it can't find either (or it's not a string) it
will pass the data through untouched.

Fixes #6944