Add support for larger pulse/alert tables
Previously pulses and alerts that used a table visualization were
limited to 3 columns and 10 rows. This commit changes the backed to
bump that to 10 columns and 20 rows. When a resultset has more data
than will fit into a 10x20 table, it will trigger a CSV attachment
that will include the full results of the query, up to 2000 results.

This commit also limits all queries to 2000 rows, similar to how queries
issued via the UI are constrained. Users wanting more than 2000 rows
use the existing export functionality.