Change how pulse table column limits are enforced
This commit differentiates between the number of columns in a
resultset and the number of columns that would actually be included in
a rendered pulse table when attempting to calculate a column
limit. Now only when the number of rendered columns (i.e. not a
description type or a details-only visualization type) exceeds the
column limit will it truncate. It will also include a CSV now if it
exceeds the column limit OR if there is a hidden column (due to it
being a description or detail-only etc).