Improve memory usage for computation job result cleanup
Previously the code was using `delete!` which will read in each of the
rows to trigger related deletes before issuing a delete of that
row. The model doesn't have any `pre-delete` actions that are needed
so this switches to a `simple-delete!` which will not cause any of the
rows to be read but rather just a delete statement to the database.

Fixes #7184