Allow for multiple hashes when constructing an app name for google [ci drivers]
For local/dev builds we include the git hash of the code in the header
information when connecting to a Google service like
BigQuery. Typically only one hash is returned when we run the shell
command `git show-ref --head --hash=7 head`. Not sure why but that can
return more than one hash separated by a line break. That line break
causes the connection to Google. It looks to be putting that in a URL
which is not an allowed character. This commit Base64 encodes that
string to ensure it's always safe to use in connecting to the Google
service.