Put together a report and database timezone test suite [ci drivers]
The goal is to make it easier to write tests with different data and
report timezone scenarios easier. The tests already have found issues
with our current code and their handling of these timezone
conversions. These tests exposed a bug in MySQL date support which is
also fixed in this commit.

Fixes #6554, fixes #6994, fixes #7393