Fix memory leak found using RedShift JDBC Metadata methods
Running sync calls several JDBC metadata methods that return
`ResultSet` objects. According to the JDBC spec, when the connection
associated to the JDBC ResultSet has been closed, it should free the
resources used by the ResultSets that were left open in that
connection. That doesn't look to be happening in RedShift. Closing the
ResutSet instance ourselves before the connection closes will ensure
that these resources are freed and doesn't allow the RedShift bug to
occur.