Change the core_user.updated_at migration
MySQL 5.5 and before doesn't support computed default values for
datetime fields. It supports them in later versions, and on timestamp
fields for 5.5, but not datetime fields, which is what we use for our
auditing columns currently. This comment just removes that default
computed value and the related not null constraint.