Switch a field values sync test from an atom to promise [ci drivers]
A call to `rescan_values` launches a future but doesn't wait for it to
finish before checking that the function was called. We have no
guarantees when the future will run. This was causing transient test
failures, depending on when that future would run. This commit
switches from checking an atom to blocking on a promise. This insures
that the future would be ran (and function would be invoked).