Add direct child collections to search results
Today searching for something in a collection only includes
cards,dashboards and pulses. If you think about collections being more
like folders, direct child collections of that collection exist in
that folder and should also be included in the results.

This commit adds support for including those direct children when
searching a collection