Handle Azure Cosmos DB connections
When attempting to connect with an Azure Cosmos DB instance,
`can-connect` for Mongo driver returns false due to a mismatch in
returned data types.

Mongo DB instances using wire protocol versions prior to 3.4 return
an an integer value (1) as the value for the "ok" field in the
response for db.stats(). Newer versions respond with a float value
(1.0). Azure Cosmos DB uses wire protocol version 3.2 by default.

This changeset casts the response value to float to ensure expected
behaviour.

Resolves: #5571