Fix transient issue with pulse search test
The test query to ensure alerts aren't returned with pulse search, and
just filtering out the id of the pulse to ensure it's not in the
results. That test can fail if there happens to be a search result
with the same id as the (missing) pulse. Adding another predicate to
ensure it's only looking for pulses.