Change the Pulse API to accept a CardRef or HybridPulseCard
The pulse GET responses include a PulseCard that isn't exactly a
PulseCard. It includes some extra information from the underlying card
useful for the FE to include in the display. When adding a card to an
existing pulse that has one or more cards already, the PUT include a
mixture of these HybridPulseCards (the existing cards for the pulse)
and CardRefs (for newly added cards). This commit changes the schema
to allow either and updates the tests accordingly.

This also adds similar flexibility to the POST so that it is
consistent with the PUT.

Fixes #7789