Allow positive numeric values for text fingerprint average lengths
Previously this only allowed double values. If the average length
happened to be something like `13.0` some Javascript implementations
trim the `.0` and when the frontend POSTs the data, it includes `13`
instead. This was causing our schema validation to break as it was
looking for a `Double`.

This commit changes the schema to look for a positive numeric value,
not just a positive double which covers both cases.