Fix nested query issue with SparkSQL driver [ci drivers]
Previously the driver was only looking one and two levels deep for the
source table. In the case of nested and binned queries, this could be
buried deeper in the MBQL query. This commit switches form an `or`
clause and a couple of `get-in`s to a `postwalk` that will find the
table however it's nested in the query.