Don't include dimension options for nested PK/FK fields [ci drivers]
There was code to remove the special_type from the field before
returning the metadata. That code was in place so that the frontend
doesn't treat that field like a real PK/FK (since it's a nested
query). That special_type was being removed before the code to added
dimension options was ran, causing it to add dimension options for
PK/FK values (which it shouldn't). This commit just separates out the
code that clears the special type so it can run after dimension
options are added.