Driver pulse table cell color from visualiation settings
This creates a new shared javascript file in
`resources/frontend_shared` that will be used by the frontend and the
backend to drive the background color of cells using conditional
formatting features the frontend has.

This uses the JDK 8 included Nashorn javascript engine to eval and
invoke the functions in that shared file.