Parameterize quoting strategy
Quoting all objects works for MySQL/MariaDB but that same setting
causes H2 and PostgreSQL to fail. H2 appears to be quoting table names
in a create table statement but not quoting that same table name when
referenced via a foreign key which fails.

This commit sets up the quote all objects strategy for MySQL/Maria but
leaves the `LEGACY` quoting strategy for H2/PostgreSQL.