Don't decrypt obviously not encrypted strings
When `MB_ENCRYPTION_SECRET_KEY` is set, we assume that all settings
have been encrypted. If a setting has not been encrypted, decryption
fails, but the original setting is returned. This causes Metabase to
behave correctly but have very noisy logging as when the decrypt fails
an error is logged.

This commit changes the `maybe-decrypt` behavior to more closely
examine the ciphertext to ensure it is something that can be
decrypted. If it looks like something that can be decrypted, but
decription fails (i.e. what if an unencrypted string is multiple of
the cipher block size) then log a warning as it's probably not an
issue, but could be.

Fixes #8151