Change the resolver to allow breakouts on nested FK related fields
Mostly this change involves resolving the nested query, then using
that resolved nested query along with the top level breakout clause to
figure out if we need to add more fields to be projected from that
inner query.

The code previously was resolving that nested query but wasn't
resolving the breakouts after and reconciling that fields list. This
also adds some tests for these nested FK scenarios.