Fix bug with MySQL date tz conversion for native queries [ci drivers]
The issue was around our usage of `du/format-date` with a built-in
date format (`:date-hour-minute-second-ms`). Those Joda Time date
formatters don't default to the system timezone, which the code
assumed. Instead it defaults to UTC. We have code in the MySQL driver
that works around a MySQL driver bug by converting the date to a
string and then using the MySQL `convert_tz()` function. That code was
converting the date from the system timezone to the report timezone,
rather than UTC, which is what the Joda Time date formatter was
returning.

Fixes #8262