Ensure we handle no git Metadata for BQ startup [ci drivers]
When connecting to a Google service, we identify ourselves with a
string that includes version, hash and branch information. If a user
downloads a Metabase tarball from Github, it won't have that Git
metadata and we will throw a NullPointerException on startup.

This commit just ensures we always have a value for that string, even
when there is no git metadata available.

Fixes #8368