core.async take/put ops in `go` blocks use single `!` functions
When using the core async `alts` and `>` functions, they should use
the single `!` version (i.e. `alts!` and `>!` instead of the `!!`
version. When using HTTP/2 and a reverse proxy, the multiplexed HTTP/2
requests get issues to Metabase not as a single connection, but as
multiple requests at the same time. The combination of those
simultaneous requests and the incorrect usage of the `core.async`
functions leads to the query responses not being delivered.

Fixes #8312