Add new flag for BigQuery to use JVM timezone [ci drivers]
This commit adds a flag to allow dates parsed from BigQuery to be
parsed in the JVM's timezone rather than UTC. By default all BigQuery
dates are in UTC but it's possible to modify the timezone via a
function invocation. This flag allows installs that have all of their
queries modifying the timezone in that way to match the actual and
expected timezones via setting the JVM's timezone.