Fix NPE from sync operations logging
This looks to have been introduced by 96b7716. Not all sync steps have
extra metadata (and thus a `log-metadata-fn` function. Unfortunately
this wasn't covered by tests as it was in the body of a `log/debug`
call. This commit fixes the problem and also pulls the logic to verify
it out of that `log/debug` call and into a separate function so that
it can be tested separately.