Add context to the IOException when creating temp files
When sending an email with an attachment, we write the attachment as a
temp file. This uses the `java.io.File/createTempFile` method. If for
some reason we're not able to write that file, maybe because we don't
have sufficient permissions, that method will throw an
IOException. The IOException doesn't include the path of the file it
was trying to create, which makes it difficult to diagnose and
fix. This commit catches that original IOException and rethrows it
with the path of the temp directory that is being used.

Fixes #8405