Specify English locale when formatting H2 SQL identifiers
Table names for H2 need to ensure that the English locale is specified
when uppercasing. Using the default locale when uppercasing can lead
to surprising results. Our `SETTING` table is a good example. If the
user has set the locale to Turkish, calling `.toUpperCase` on
"setting" will result in `SETTİNG`, which is not the same as `SETTING`
and will cause an error.

Fixes #8615, fixes #8565