Ensure we don't try to hydrate a nil pulse
It's possible if the pulse is archived that we'll try to hydrate it
and fail a schema validation. This will fail to send a pulse we
weren't going to send anyway, but makes a lot of noise in the log that
won't go away. This commit avoids that hydration if we don't have an
unarchived pulse.

Fixes #8581