Added BQ style test data setup.
Still not working. The most recent error is one where Snowflake is
expecting us to declare which database we're using. The error
references including a `USE DATABASE` statement before querying. It
looks like we were trying to solve that with a connection parameter,
but it doesn't appear to be working.