Remove bad timezone test [ci drivers]
Originally this test was attempting to recreate an issue where the
report-timezone was incorrectly applied to a database that didn't
support it. I thought the test being removed in this commit was doing
that but when I examine the resulting SQL more closely, it was not,
rather it appeared as thought it did, but due to some odd truncation
we were applying to dates, it wasn't.