Fix for a transient test issue on Snowflake [ci drivers]
Specifically two tests with a between date filter are failing. Only
sometimes and only when running the full test suite. The guess right
now is it's grabbing an old connection with a previous test
report-timezone already preset on the connection. This commit flushes
the connection pool before running the test (only on Snowflake).