Change query result hashing to convert integers to floats first
When the FE returns the query result metadata, any integer values will
be converted to a floating point number. Those doubles hash different
than the original integers which causes the hashes not to match and
the query to be reran when saving the question. This commit just
converts the integer to a double before computing the hash.

Fixes #8824