Fix test checksum test failure on MySQL [ci drivers]
The newly added checksum tests fail when running on MySQL in
Circle. The MySQL tests include a secret key for encryption which are
used when computing the checksum. Because of that encryption, the
checksum string is different every time (it must be decrypted before
compared). The test isn't covering the encryption code but is only
concerned with ensure the datastructure hashes consitently. This
commit uses a `with-redefs` to set that secret key to `nil` so that
encryption won't be used on those two tests.