Only include a CSV attachment for pulses when XLS is not selected
Previously we would always include a CSV attachment when there were
more rows/columns than would be included in the table. This ended up
being redundant for users that would prefer an XLS attachment and have
indicated that the pulse should include an XLS attachment. This commit
will only not attach a CSV if the user has select an XLS attachment
for that pulse.

Fixes #7302